Move LLM facade tests to llm package
