+++ pkg/agentcompose/coverage_shape_workflows_test.go
@@ -38,7 +38,6 @@ func TestIntegrationWebhookWorkspaceAndLoaderWorkflow(t *testing.T) {
 	testAgentDefinitionValidationAndProtoMapping(t)
 	testAgentDefinitionCreateSession(t)
 	testDeleteAgentDefinitionStopsSessionsAndKeepsDeletedInList(t)
-	testRuntimeLLMFacadeCoverageWorkflows(t)
 	testWebhookIntegrationEventDispatchRunsMatchingLoader(t)
 }
 
@@ -78,28 +77,6 @@ func TestE2EWebhookWorkspaceAndLoaderWorkflow(t *testing.T) {
 	testAgentDefinitionValidationAndProtoMapping(t)
 	testAgentDefinitionCreateSession(t)
 	testDeleteAgentDefinitionStopsSessionsAndKeepsDeletedInList(t)
-	testRuntimeLLMFacadeCoverageWorkflows(t)
 	testServiceReconcilePersistedSessionsMarksStaleProjectRunsFailed(t)
 	testWebhookIntegrationEventDispatchRunsMatchingLoader(t)
-}
-
-func testRuntimeLLMFacadeCoverageWorkflows(t *testing.T) {
-	t.Helper()
-	TestRuntimeLLMUseGenericResponsesTextPartsRequiresExplicitProviderFlag(t)
-	TestRuntimeLLMFacadeForwardsWithSessionToken(t)
-	TestRuntimeLLMFacadeFlushesSSEResponses(t)
-	TestRuntimeLLMAnthropicFacadeForwardsWithSessionToken(t)
-	TestRuntimeLLMOpenAIResponsesFacadeBridgesToAnthropicProvider(t)
-	TestRuntimeLLMAnthropicFacadeBridgesToOpenAIResponsesProvider(t)
-	TestAnthropicProviderCanBootstrapFromGenericLLMAPIKey(t)
-	TestRuntimeLLMTargetDoesNotTreatGenericOpenAIEndpointAsAnthropic(t)
-	TestManagedRuntimeEnvMapKeepsFacadeKeyAliases(t)
-	TestEnsureSessionLLMFacadeConfigUsesRequestedModel(t)
-	TestEnsureSessionOpenCodeCustomProviderWritesConfig(t)
-	TestEnsureSessionOpenCodeOpenAIWritesRequestedModelConfig(t)
-	TestEnsureSessionOpenCodeAnthropicUsesFacadeEnv(t)
-	TestEnsureSessionLLMFacadeConfigBootstrapsFromSessionEnvProvider(t)
-	TestCodexFacadeCanUseAnthropicOnlySessionEnvProvider(t)
-	TestServiceGenerateLLMChatCompletionsProtocol(t)
-	testRuntimeLLMFacadeRoundTrips(t)
 }
