+++ pkg/llm/model_alias.go
@@ -6,13 +6,17 @@ import (
 )
 
 const (
+	VMStatusRunning = model.VMStatusRunning
 	VMStatusStopped = model.VMStatusStopped
 	VMStatusFailed  = model.VMStatusFailed
+
+	SessionTypeManual = model.SessionTypeManual
 )
 
 type ConfigStore = storage.ConfigStore
 type Store = storage.Store
 type Session = model.Session
+type SessionSummary = model.SessionSummary
 type SessionEnvVar = model.SessionEnvVar
 type LLMProvider = model.LLMProvider
 type LLMModel = model.LLMModel
